adapt RegularCard from creativeTim
this adapts the regularCard component from the creative tim material
template. It also adds the necessary JSS files and the dependencies.
The problem with it is though that I couldn't find out how to get
"withStyles" to work.
There is some documentation here: https://material-ui-next.com/guides/typescript/#usage-of-withstyles
but even with that I couldn't get it to work.
Help is definitely wanted and needed to finish this.